metric calculation
This is the basic calculation of the new metric. I have been unable to find a database that stores records for BQR-S (bequeathed runners scored, or, in simpler terms, runners that scored between two, or three, different pitchers. Once I find such a database I'll be able to continue on with the code. The plan is to first have a solid foundation that can return values for starting pitchers, then I'll start working on fine tuning it so that it will work on relievers, and even when 3 pitchers contribute to a runner scoring. I'm still a relative beginner in programming, so the code isn't much yet, but I'm a believer in this statistic and will continue working on it.